+++ libs/Core/Router.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Lib;
+namespace libs\Core;
 
 class Router
 {
@@ -172,7 +172,7 @@ class Router
             if (!self::$error_callback) {
                 self::$error_callback = function () {
                     header($_SERVER['SERVER_PROTOCOL'] . " 404 Not Found");
-                    include './resources/404.php';
+                    Message::send(404, [], 'The requested URL ' . $_SERVER['REQUEST_URI'] . ' was not found on this server.');
                 };
             } else {
                 if (is_string(self::$error_callback)) {