+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.3.0 - 2023.04.07
+
+- Upgraded DB class
+
+- Modified the file directory structure (libs is placed in the root directory)
+
+- Change the apiout class to message, and redefine the output format
+
 ## 0.2.0
 
 - Risen init